app/requirements.txt
diffusers[torch]==0.27.2
accelerate
torch==2.2.2
transformers==4.40.0
xformers==0.0.25.post1

invisible_watermark # To help viewers identify the images as machine-generated.
onnxruntime==1.17.3 # ONNX Runtime uses the following optimizations to speed up Stable Diffusion in CUDA.
safetensors==0.4.3 # To store tensors safely.
controlnet_aux

Pillow
pillow-avif-plugin # To save images in AVIF format.
pyyaml # To read the configuration file by written YAML.
peft